+++ client/config/modules.js
