+++ client/config/paths.js
