+++ client/scripts/start.js
