+++ client/src/components/FirstNavbar/More/More.js
@@ -6,13 +6,7 @@ const More = () => (
     <div className={classes["navbar-1__content__MoreSection__More"]}>
       <span className={classes["navbar-1__content__MoreSection__More__myMore"]}>
         More
-        <svg
-          className={classes["icon icon-cheveron-down"]}
-          width="9"
-          height="20"
-          viewBox="5 3 10 1"
-          fill="#fff"
-        >
+        <svg className={classes["icon icon-cheveron-down"]} width="9" height="20" viewBox="5 3 10 1" fill="#fff">
           <use xlinkHref="#icon-cheveron-down"></use>
           <path
             d="M9.293 12.95l0.707 0.707 5.657-5.657-1.414-1.414-4.243
