+++ client/src/containers/StockManager/StockManager.js
